Do not run OSM data update script in endless loop
This can be achieved with Systemd as well using a timer. It has the
advantage that we can just disable the timer to safely stop updates
after a single run is completed. Currently, we have to wait with sending
the signal to stop the service until it is in sleep mode.