map upgrade: add more options to script
Allow to skip style updates completely, which is useful e.g. to get updated
legend or JOSM styles. Allow to specify the bbox in style mode as well, which
is useful when e.g. the signal style only added support for a new system used
only in a certain area.