Add knallfrosch as tileserver, make role more flexible
knallfrosch does not have a /nvme directory because it makes no sense on
a host with all disks being NVMe drives.

This commit improves handling of the transient Systemd unit for OSM data
import. You should now be able to just run Ansible as often as you want
until the import is complete. Once the import is finished, all remaining
tileserver setup tasks will be executed.