Fix Bash syntax
Test -v should be followed by the variable name without $, otherwise the
variable is expanded first.

This bug made the import use the planet_osm_nodes table in the database
but the diff updates use the flatnodes file. As a consequence, diff
updates removed updated ways from the database if their nodes have not
been changed since the first import.