Fix Debian version check
Prior to this fix, to wrong path to the Tirex/renderd socket was set. It
was not noticed until today because there were some changes on the
server which were not in Ansible and I did not notice that when I
deployed commit a0d3267adde1d81e376612097eb4e2fedac2021e.